TIP-977: Move CommandExecutor to Tool
