+++ tests/back/Platform/Specification/Bundle/InstallerBundle/Event/InstallerEventSpec.php
@@ -2,8 +2,8 @@
 
 namespace Specification\Akeneo\Platform\Bundle\InstallerBundle\Event;
 
-use Akeneo\Platform\Bundle\InstallerBundle\CommandExecutor;
 use Akeneo\Platform\Bundle\InstallerBundle\Event\InstallerEvent;
+use Akeneo\Tool\Component\Console\CommandExecutor;
 use PhpSpec\ObjectBehavior;
 use Symfony\Component\EventDispatcher\GenericEvent;
